[SYCL] Add Unified Runtime Plugin (#7521)
Add new pi_unified_plugin that implements only piPlatformsGet (via Unified Runtime zerPlatformGet)
The next step is to add a new backend to SYCL and actually start calling to pi_unified_plugin
