+++ sycl/plugins/level_zero/pi_level_zero.hpp
@@ -42,13 +42,13 @@
 #include <unordered_set>
 #include <vector>
 
-#include <level_zero/ze_api.h>
-#include <level_zero/zes_api.h>
 #include <sycl/detail/iostream_proxy.hpp>
+#include <ze_api.h>
+#include <zes_api.h>
 
 // Share code between this PI L0 Plugin and UR L0 Adapter
-#include <adapters/level_zero/ur_level_zero.hpp>
 #include <pi2ur.hpp>
+#include <ur/adapters/level_zero/ur_level_zero.hpp>
 
 #include "usm_allocator.hpp"
 
@@ -58,7 +58,7 @@ template <class To, class From> To pi_cast(From Value) {
   return (To)(Value);
 }
 
-template <> uint32_t pi_cast(uint64_t Value) {
+template <> uint32_t inline pi_cast(uint64_t Value) {
   // Cast value and check that we don't lose any information.
   uint32_t CastedValue = (uint32_t)(Value);
   assert((uint64_t)CastedValue == Value);
@@ -180,13 +180,8 @@ struct MemAllocRecord : _pi_object {
 // Define the types that are opaque in pi.h in a manner suitabale for Level Zero
 // plugin
 
-struct _pi_platform : public _ur_level_zero_platform {
-  _pi_platform(ze_driver_handle_t Driver) : _ur_level_zero_platform{Driver} {}
-
-  // Performs initialization of a newly constructed PI platform.
-  pi_result initialize() {
-    return ur2piResult(_ur_level_zero_platform::initialize());
-  }
+struct _pi_platform : public _ur_platform_handle_t {
+  using _ur_platform_handle_t::_ur_platform_handle_t;
 
   // Cache pi_devices for reuse
   std::vector<std::unique_ptr<_pi_device>> PiDevicesCache;
