+++ sycl/plugins/level_zero/ur_bindings.hpp
@@ -1,0 +1,17 @@
+//===------ ur_bindings.hpp - Complete definitions of UR handles -------==//
+//
+// Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+//
+//===------------------------------------------------------------------===//
+#pragma once
+
+#include "pi_level_zero.hpp"
+#include <zer_api.h>
+
+// Make the Unified Runtime handles definition complete.
+// This is used in various "create" API where new handles are allocated.
+struct _zer_platform_handle_t : public _pi_platform {
+  using _pi_platform::_pi_platform;
+};
