+++ sycl/plugins/unified_runtime/pi_unified_runtime.hpp
@@ -1,10 +1,8 @@
-//===---------------- pi2ur.cpp - PI API to UR API  --------------------==//
+//===--- pi_unified_runtime.hpp - Unified Runtime PI Plugin  -----------==//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 //
 //===------------------------------------------------------------------===//
-
-// This thin layer performs conversion from PI API to Unified Runtime API
-// TODO: remove when SYCL RT is changed to talk in UR directly
+#pragma once
