+++ sycl/test/basic_tests/interop-backend-traits.cpp
@@ -1,5 +1,5 @@
 // RUN: %clangxx -fsycl -DUSE_OPENCL %s
-// RUN: %clangxx -fsycl -DUSE_L0 %s
+// RUN: %clangxx %fsycl-host-only -DUSE_L0 %s
 // RUN: %clangxx -fsycl -DUSE_CUDA %s
 // RUN: %clangxx -fsycl -DUSE_HIP %s
 // RUN: %clangxx -fsycl -DUSE_CUDA_EXPERIMENTAL %s
@@ -13,7 +13,7 @@ constexpr auto Backend = sycl::backend::opencl;
 #endif
 
 #ifdef USE_L0
-#include <level_zero/ze_api.h>
+#include <ze_api.h>
 
 #include <sycl/ext/oneapi/backend/level_zero.hpp>
