+++ sycl/test/lit.site.cfg.py.in
@@ -12,6 +12,7 @@ config.sycl_libs_dir = lit_config.params.get('SYCL_LIBS_DIR', "@LLVM_LIBS_DIR@")
 config.target_triple = "@LLVM_TARGET_TRIPLE@"
 config.host_triple = "@LLVM_HOST_TRIPLE@"
 config.opencl_libs_dir = os.path.dirname("@OpenCL_LIBRARIES@")
+config.level_zero_include_dir = "@LEVEL_ZERO_INCLUDE_DIR@"
 config.opencl_include_dir = "@OpenCL_INCLUDE_DIR@"
 config.cuda_toolkit_include = "@CUDA_TOOLKIT_INCLUDE@"
 config.sycl_tools_src_dir = "@SYCL_TOOLS_SRC_DIR@"
