+++ sycl/tools/sycl-trace/CMakeLists.txt
@@ -45,13 +45,13 @@ add_custom_target(pi-pretty-printers
 if ("level_zero" IN_LIST SYCL_ENABLE_PLUGINS)
   add_dependencies(sycl_pi_trace_collector pi_level_zero)
 
-  target_link_libraries(sycl_pi_trace_collector PRIVATE LevelZeroLoader::Headers)
+  target_link_libraries(sycl_pi_trace_collector PRIVATE LevelZeroLoader-Headers)
   target_compile_definitions(sycl_pi_trace_collector PRIVATE SYCL_HAS_LEVEL_ZERO)
 
   add_custom_target(ze-pretty-printers
     COMMAND ${Python3_EXECUTABLE}
     ${CMAKE_CURRENT_SOURCE_DIR}/generate_ze_pretty_printers.py
-    ${SYCL_INCLUDE_BUILD_DIR}/sycl/level_zero/ze_api.h
+    ${LEVEL_ZERO_INCLUDE_DIR}/ze_api.h
     DEPENDS pi_level_zero
     BYPRODUCTS
       ${CMAKE_CURRENT_BINARY_DIR}/ze_printers.def
